backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
pandas
openpyxl
gunicorn
psycopg2-binary
python-dotenv
